server/requirements.txt
flask
paho-mqtt
influxdb-client
requests
python-dotenv
pyyaml